requirements.txt
openai
chromadb
tiktoken
jieba
streamlit
python-docx
PyPDF2
gensim 